Working solution with context managers
