+++ esd_services_api_client/nexus/algorithms/minimalistic.py
@@ -1,7 +1,5 @@
 from abc import ABC
 
-from pandas import DataFrame as PandasDataFrame
-
 from esd_services_api_client.nexus.algorithms._baseline_algorithm import (
     BaselineAlgorithm,
 )
